% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/branje.izpita.R
\name{branje.izpita}
\alias{branje.izpita}
\title{branje.izpita}
\usage{
branje.izpita(
 
    test = image_read("https://raw.githubusercontent.com/borbregant/Rizpit/master/tests/test.png")
)
}
\arguments{
\item{test}{skeniran izpit v obliki pnd}
}
\description{
Funkcija izvozi ime kandidata in njegove tocke na skeniranem izpitu
}
\examples{
branje.izpita()
}
\keyword{branje}
\keyword{ocr,}
\keyword{sken,}
