% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcijagoljuf.R
\name{funkcijagoljuf}
\alias{funkcijagoljuf}
\title{funkcijagoljuf}
\usage{
funkcijagoljuf(myData)
}
\arguments{
\item{myData}{input tabela ki sprejme nujno v obliki ...}
}
\description{
Funkcija preveri kandidate za goljufe
}
\examples{
funkcijagoljuf()
}
\keyword{goljuf}
