% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcijastatistika.R
\name{funkcijastatistika}
\alias{funkcijastatistika}
\title{funkcijastatistika}
\usage{
funkcijastatistika(myData, maxtock)
}
\arguments{
\item{myData}{input tabela ki sprejme nujno v obliki IME|TOCKE1|...|TOCKEn|}

\item{maxtock}{maksimalno mozno stevilo dosezenih tock}
}
\description{
Funkcija izvozi txt in pdf statistike izpita myData
}
\examples{
funkcijastatistika()
}
\keyword{statistika}
