% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocena.R
\name{ocena}
\alias{ocena}
\title{ocena}
\usage{
ocena(izpit, skupnotock)
}
\arguments{
\item{skupnotock}{stevilo max tock, ki jih je bilo mozno doseci na izpitu}

\item{myData}{input tabela ki sprejme nujno v obliki VPISNA|TOCKE1|...|TOCKEn|}
}
\description{
Funkcija iz dataframea izpita izlusci ocene posameznih kandidatov
}
\examples{
ocena()
}
\keyword{ocena}
