% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urejanjetabel.R
\name{urejanjetabel}
\alias{urejanjetabel}
\title{urejanjetabel}
\usage{
urejanjetabel(
  kolokvij1,
  kolokvij2 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  kolokvij3 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  kolokvij4 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  kolokvij5 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  kolokvij6 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  kolokvij7 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  kolokvij8 = data.frame(IME = c(0), PRVA = c(0), DRUGA = c(0), TRETJA = c(0)),
  skupnotock,
  zapisi = FALSE
)
}
\arguments{
\item{skupnotock}{skupno stevilo tock kolokvijev, da namesto tock v skupno tabelo vpise procent}

\item{zapisi}{skupno tabelo zapise v obliki dataframe-a. zaenkrat deluje le na mojem lokalnem racunalniku}

\item{kolokviji}{input tabela ki sprejme nujno v obliki IME|TOCKE1|...|TOCKEn|}
}
\description{
Funkcija iz dataframeov posameznih izpitov !najvec8! sestavi skupno tabelo, ki prikaze le kandidata in skupno tock
Lahko podamo argument skupnotock=, ki namesto skupnega stevila tock priredi oceno vsakemu kandidatu
}
\examples{
urejanjetabel()
}
\keyword{ocena}
